all: use new nodejs loadEnvFile
