+++ packages/hop-node-core/src/config/config.ts
@@ -7,10 +7,14 @@ import {
 import { type Metadata, metadata as coreMetadata } from '@hop-protocol/sdk/metadata'
 import { type Networks, networks as coreNetworks } from '@hop-protocol/sdk/networks'
 import { execSync } from 'node:child_process'
-import { loadEnv } from './loadEnvFile.js'
-import { normalizeEnvVarNumber } from './utils/normalizeEnvVarNumber.js'
+import { getEnvFilePath, normalizeEnvVarNumber } from './utils.js'
+import { loadEnvFile } from 'node:process'
 
-loadEnv()
+
+const envFilePath = getEnvFilePath()
+if (envFilePath) {
+  loadEnvFile(envFilePath)
+}
 
 const bonderPrivateKey = process.env.BONDER_PRIVATE_KEY
 export const setLatestNonceOnStart = process.env.SET_LATEST_NONCE_ON_START
