+++ packages/hop-node-core/src/config/index.ts
@@ -1,3 +1,2 @@
 export * from './config.js'
-export * from './utils/index.js'
-export * from './loadEnvFile.js'
+export * from './utils.js'
