--- packages/hop-node-core/src/config/utils/normalizeEnvVarArray.ts
@@ -1,3 +1,0 @@
-export function normalizeEnvVarArray (value?: string): string[] {
-  return (value ?? '').split(',').map(x => x.trim()).filter(x => x)
-}
