--- packages/hop-node-core/src/config/utils/normalizeEnvVarNumber.ts
@@ -1,5 +1,0 @@
-export function normalizeEnvVarNumber (value?: string): number | undefined {
-  if (value !== undefined) {
-    return Number(value.toString())
-  }
-}
