+++ packages/v2-explorer-backend/package.json
@@ -40,7 +40,6 @@
     "async-mutex": "0.3.2",
     "commander": "6.2.1",
     "cors": "2.8.5",
-    "dotenv": "16.0.0",
     "ethers": "5.7.1",
     "express": "4.17.3",
     "express-rate-limit": "7.2.0",
