+++ packages/v2-explorer-backend/src/config/config.ts
@@ -1,9 +1,14 @@
  
-import dotenv from 'dotenv'
 import os from 'node:os'
-import { TextDecoder, TextEncoder } from 'node:util'
+import { TextEncoder } from 'node:util'
+import { getEnvFilePath } from '#utils/getEnvFilePath.js'
+import { loadEnvFile } from 'node:process'
+
 global.TextEncoder = TextEncoder
-dotenv.config()
+const envFilePath = getEnvFilePath()
+if (envFilePath) {
+  loadEnvFile(envFilePath)
+}
 
 export const dbPath = process.env.DB_PATH ?? '/tmp/tempdb'
 export const privateKey = process.env.PRIVATE_KEY
